requirements.txt
asgiref==3.2.3
astroid==2.1.0
autopep8==1.4.3
Click==7.0
Django==3.0.3
django-tables2==2.2.1
Flask==1.1.1
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pylint==2.2.2
pylint-django==2.0.6
pylint-plugin-utils==0.5
python-slugify==1.1.4
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
Unidecode==1.0.23
Werkzeug==0.16.1
wrapt==1.11.1
